Build splash page and split up CSS.
This introduces the first version of a splash page with a photo,
contact information, and a list of articles, currently with placeholder
articles only.

Additionally, CSS formatting links, applying fonts and the like, that is
likely to be shared between multiple style sheets has been split into a
base.css that can be imported in different layouts' style sheets.