Make category descriptions visible to Jekyll SEO
This is very hacky.

Ideally, we could store category descriptions in site.data.categories,
and set the page.description variable based on that. Unfortunately, this
does not seem to be possible.

Instead, we have to store the description in the pages themselves. This
introduces the new problem of how to list the subcategories in the
livereads page. Ultimately, my workaround was to go through all the
site.pages and identify the category index page by URL, then extract the
description from there.

It's ugly, but it works.