Implement 'choice box' include for quiz questions
This allows an unordered list to be transformed into a list of clickable options, with the ability to specify a correct answer. When an option is clicked, it will either be given a tick or a cross and the other answers greyed out; in the latter case the correct answer will be highlighted and marked with an arrow.